Feat: Add support for redeployment of an existing environment (#62)
* Update: Add support for environment rebuild in upload API

Modified the upload API to support updating an existing environment of a property and added functionality for triggering a redeployment of an existing environment. Introduced a new boolean attribute 'rebuild' in the form data. When set to true, it initiates a redeployment process while keeping all other form attributes and processes unchanged. This commit includes all related changes.

* Refactor: Change `rebuildEnvironment` field type to String and add parsing logic

* Fix: Address "spec is immutable" error for bound claims